Rename build-page-targeting to page-targeting
