+++ src/lib/page-targeting.spec.ts
@@ -1,7 +1,7 @@
 import type { ConsentState } from '@guardian/libs';
 import { buildPageTargeting as buildPageTargeting_ } from '../core/targeting/build-page-targeting';
-import { getPageTargeting } from './build-page-targeting';
 import { isUserLoggedInOktaRefactor } from './identity/api';
+import { getPageTargeting } from './page-targeting';
 
 const buildPageTargeting = buildPageTargeting_ as jest.MockedFunction<
 	typeof buildPageTargeting_
